fix warnings

+++ vp/src/core/common/trace.h
@@ -87,10 +87,10 @@ struct ScoreParams {
 	Opcode::Mapping tree; 
 	uint64_t weight; 
 	uint32_t length; 
-	float dep_score;
-	int32_t num_children;
-	int32_t inputs;
-	int32_t outputs; 
+	double dep_score;
+	uint32_t num_children;
+	uint32_t inputs;
+	uint32_t outputs; 
 	float score_multiplier; 
 	float score_bonus; 
 // uint32_t num_pcs;
@@ -234,11 +234,11 @@ struct CsvParams {
 };
 
 struct PathExtensionParams {
-    uint32_t length;
+    int32_t length;
     float score_bonus;
     float score_multiplier;
     int32_t tree_id;
-    int force_extension_depth;
+    int32_t force_extension_depth;
     Opcode::Mapping force_instruction;
 	std::function <float(const ScoreParams)> score_function;
 };
@@ -247,7 +247,7 @@ struct BranchingPoint
 {
 	uint8_t depth = 0;
 	Opcode::Mapping instruction = Opcode::UNDEF;
-	uint64_t weight = 0;
+	int64_t weight = 0;
 	double ratio = 0.0;
 	InstructionNode* starting_point;
 };
@@ -699,7 +699,7 @@ class InstructionNode{
 
 		//find a point in an existing sequence with the highest ratio between branch taken in the original sequence 
 		// and another possible branch not taken, which would lead to a different sequence 
-		virtual std::vector<BranchingPoint> find_variant_branch(Path path, uint depth){
+		virtual std::vector<BranchingPoint> find_variant_branch(Path path, uint8_t depth){
 			return {}; //no possible branching points for leaf nodes
 		}
 
@@ -747,7 +747,7 @@ class InstructionNodeR : virtual public InstructionNode{
 		std::map<uint64_t, int> get_pc() override;
 
 		std::vector<PathNode> path_to_path_nodes(Path path, uint depth) override; 
-		std::vector<BranchingPoint> find_variant_branch(Path path, uint depth) override;
+		std::vector<BranchingPoint> find_variant_branch(Path path, uint8_t depth) override;
 		int prune_tree(uint64_t weight_threshold, uint8_t depth) override;
 		NODE_TYPE get_node_type() override;
 };
